scales required but not exported
percent() is part of the scales package but wasn't specified as an import. percent() has been superseded by label_percent(), which returns a function rather than a vector. To work with current data, required to use map_chr from the purrr package to map the function to the ev.rate column.