added parameters source and branch
branch allows the user to download the version of functions from a specific branch in the sirfunction GitHub repo. source determines whether the user should use desk review functions that come with the user's version of sirfunctions or if it should use the functions located in the R folder of the desk review of a  country.